Renamed Table to FreeFloorPlan and moved FreeFloorPlan and Table to presentational
