+++ client/src/components/Presentational/FreeFloorPlan/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { MainContainer } from '../../global-styles/styledComponents';
+import { MainContainer } from '../../../global-styles/styledComponents';
 
 export const Container = styled(MainContainer)`
   display: flex;
