+++ client/src/components/Presentational/Table/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { Boxes } from '../../global-styles/styledComponents';
+import { Boxes } from '../../../global-styles/styledComponents';
 
 export const TableBoxes = styled(Boxes)`
   /* Style modifications go here */
